+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Timestepping: add `reparametrise_arclength` option.
+
 ## [0.34.13] - 2026-07-24
 
 ### Fixed