+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   The full velocity, which was always computed at the beginning of a timestep,
   is not needed by Strang schemes, which require only the "fast" term at that point.
 
+### Fixed
+
+- Fix filament plots with recent Makie versions.
+
 ## [0.34.13] - 2026-07-24
 
 ### Fixed