+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.34.14] - 2026-09-10
+
 ### Added
 
 - Timestepping: add `reparametrise_arclength` option.